DF-5220 fix review comment
Former-commit-id: df8f334a75a7db327f107ed59a49200756a07a62

+++ Sources/Internal/Platform/TemplateAndroid/Java/com/dava/framework/JNIDeviceInfo.java
@@ -287,8 +287,8 @@ public class JNIDeviceInfo {
 			    free = statFs.getAvailableBlocksLong() * statFs.getBlockSizeLong();
 			} else
 			{
-			    capacity = (long)statFs.getBlockCount() * (long)statFs.getBlockSize();
-			    free = (long)statFs.getAvailableBlocks() * (long)statFs.getBlockSize();
+			    capacity = statFs.getBlockCount() * statFs.getBlockSize();
+			    free = statFs.getAvailableBlocks() * statFs.getBlockSize();
 			}
 
             boolean isEmulated = Environment.isExternalStorageEmulated();
